+++ packages/vertex-sdk/src/client.ts
@@ -86,12 +86,21 @@ export class AnthropicVertex extends BaseAnthropic {
       );
     }
 
+    if (!baseURL) {
+      switch (region) {
+        case 'global':
+          baseURL = 'https://aiplatform.googleapis.com/v1';
+          break;
+        case 'us':
+          baseURL = 'https://aiplatform.us.rep.googleapis.com/v1';
+          break;
+        default:
+          baseURL = `https://${region}-aiplatform.googleapis.com/v1`;
+      }
+    }
+
     super({
-      baseURL:
-        baseURL ||
-        (region === 'global' ?
-          'https://aiplatform.googleapis.com/v1'
-        : `https://${region}-aiplatform.googleapis.com/v1`),
+      baseURL,
       ...opts,
     });