+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.83.0 (2026-04-03)
+
+Full Changelog: [sdk-v0.82.0...sdk-v0.83.0](https://github.com/anthropics/anthropic-sdk-typescript/compare/sdk-v0.82.0...sdk-v0.83.0)
+
+### Features
+
+* **vertex:** add support for US multi-region endpoint ([5e5aea7](https://github.com/anthropics/anthropic-sdk-typescript/commit/5e5aea72e8af1d0cd0af7770f7c98a716fa24629))
+* **vertex:** add support for US multi-region endpoint ([0de0e98](https://github.com/anthropics/anthropic-sdk-typescript/commit/0de0e98aace311d7f6b5617b597d73324de73e2f))
+
+
+### Bug Fixes
+
+* **client:** dont upload aws artifact ([#844](https://github.com/anthropics/anthropic-sdk-typescript/issues/844)) ([d1a31fc](https://github.com/anthropics/anthropic-sdk-typescript/commit/d1a31fcd591d61d3bd2a4cb1b0b5cfcf2f66f5fe))
+
+
+### Chores
+
+* **client:** deprecate client-side compaction helpers ([1926e87](https://github.com/anthropics/anthropic-sdk-typescript/commit/1926e870393b653ac87c4fb9e521a8a44786ab49))
+* **client:** internal updates ([3d64763](https://github.com/anthropics/anthropic-sdk-typescript/commit/3d6476315480508cff8462a5b4523944579dbd32))
+
 ## 0.82.0 (2026-04-01)
 
 Full Changelog: [sdk-v0.81.0...sdk-v0.82.0](https://github.com/anthropics/anthropic-sdk-typescript/compare/sdk-v0.81.0...sdk-v0.82.0)